aider: Change the syntax in `timekeeper.sdPlugin\actions\pauseall\property-inspector\inspector.js` to use `require` statements instead of `/// <reference path="..." />` statements.
# Aider chat conversation:

USER: should we change the reference in app.js to this /// <reference path= instead of the other require statement? which is better

ASSISTANT: The `/// <reference path="..." />` syntax is a TypeScript way of adding type definitions from another file. It's not standard JavaScript and won't work in a regular JavaScript context.

The `require` statement, on the other hand, is a part of the CommonJS module system, which is used in Node.js for importing modules.

In your case, since you're working with JavaScript and not TypeScript, you should use the `require` statement to import the `Action` class from the `StreamDeckSDK` module.

So, the existing code in `timekeeper.sdPlugin\app.js`:

```javascript
const { Action } = require('./libs/js');
```

is correct and should not be changed to a `/// <reference path="..." />` statement.

USER: what about any of the other files we have created? should we change the syntax in those? like the `timekeeper.sdPlugin\actions\pauseall\property-inspector\inspector.js`

ASSISTANT: The `/// <reference path="..." />` syntax is specific to TypeScript and is used for adding type definitions from another file. It's not standard JavaScript and won't work in a regular JavaScript context.

The `require` statement, on the other hand, is a part of the CommonJS module system, which is used in Node.js for importing modules.

In your case, since you're working with JavaScript and not TypeScript, you should use the `require` statement to import modules in your JavaScript files.

So, the existing code in `timekeeper.sdPlugin\actions\pauseall\property-inspector\inspector.js`:

```javascript
/// <reference path="../../../libs/js/property-inspector.js" />
/// <reference path="../../../libs/js/utils.js" />
```

should be changed to:

```javascript
const PropertyInspector = require('../../../libs/js/property-inspector.js');
const Utils = require('../../../libs/js/utils.js');
```

Here is the *edit block* for this change:

```javascript
timekeeper.sdPlugin\actions\pauseall\property-inspector\inspector.js
<<<<<<< ORIGINAL
/// <reference path="../../../libs/js/property-inspector.js" />
/// <reference path="../../../libs/js/utils.js" />
=======
const PropertyInspector = require('../../../libs/js/property-inspector.js');
const Utils = require('../../../libs/js/utils.js');
>>>>>>> UPDATED
```

Please review and apply this change if it looks good to you.