aider: Refactor the `writeToLogFile` function to format the date and time based on the `timeFormat` variable.
# Aider chat conversation:

USER: is the time format property being used in the log file?

ASSISTANT: Looking at the `app.js` file, it appears that the `timeFormat` variable is being set based on the user's input, but it is not being used when writing to the log file in the `writeToLogFile` function. The date and time are currently being formatted as `YYYY-MM-DD` and are not using the `timeFormat` variable.

To correct this, we need to modify the `writeToLogFile` function to format the date and time based on the `timeFormat` variable. Here is the edit block for that change:

```javascript
com.timekeeper.sdPlugin\app.js
<<<<<<< HEAD
    const date = new Date();
    const dateString = `${date.getFullYear()}-${
      date.getMonth() + 1
    }-${date.getDate()}`;
=======
    const moment = require('moment');
    const date = moment().format(timeFormat);
>>>>>>> updated
```

This change will format the date and time according to the `timeFormat` variable, which can be `hh:mm:ss`, `h hours, m minutes, s seconds`, or `yyyy-mm-dd hh:mm:ss` based on the user's selection. Please note that this change requires the `moment` library, so you will need to install it if it's not already included in your project.